refactor: extract OverlaySettings to src/server/shared to eliminate duplication
- Create src/server/shared/overlaySettings.ts as the single canonical definition
- storage.ts re-exports from shared (keeps public API stable for settings.ts)
- SettingsContext.tsx imports from shared instead of duplicating
- tsconfig.server.json already covers src/server/**/* so no config change needed
- Dockerfile CMD path unchanged (dist-server/index.js) since rootDir is still ./src/server
